# frontend/src/types/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "werewolf-agent-sim server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/games": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Instant start: solo human (role choice) or LLM-only",
        "operationId": "create_game_games_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewSoloGame"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "Status snapshot",
        "operationId": "game_status_games__game_id__get",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/key": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Resume a key-funded game after a restart by supplying the key again",
        "description": "Any seat holder may fund the resume; spectators may not. The key is tried on the\nprovider first (422 with its complaint), then the game continues from its checkpoint.\n409 when the game is not waiting for a key.",
        "operationId": "fund_game_games__game_id__key_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundGame"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/turns": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Submit the human seat's action",
        "operationId": "submit_turn_games__game_id__turns_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/draft": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Draft the seat's line from rough notes",
        "description": "The seat's agent phrases the notes into one line in the player's voice, for the\ndiscussion turn the seat owes. Nothing enters the game: the player sends the line\nwith POST /turns, edited or not, or types their own. 409 when the seat owes no\ndiscussion turn or has used this turn's drafts; 503 when the model could not answer.",
        "operationId": "draft_turn_games__game_id__draft_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/events": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "SSE event stream",
        "description": "Stream entitled durable events plus public ephemeral pacing frames.\n\n``last_seq`` is the cursor frozen into the original URL. On automatic browser\nreconnection, ``Last-Event-ID`` carries the live cursor and therefore wins when\nvalid. Pacing frames have no ID, so they never advance that durable cursor.",
        "operationId": "event_stream_games__game_id__events_get",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          },
          {
            "name": "last_seq",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Last Seq"
            }
          },
          {
            "name": "last-event-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DraftRequest": {
        "properties": {
          "notes": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "notes"
        ],
        "title": "DraftRequest",
        "description": "POST /games/{id}/draft body: rough notes for the line the seat is about to say."
      },
      "DraftResponse": {
        "properties": {
          "draft": {
            "type": "string",
            "title": "Draft"
          },
          "drafts_left": {
            "type": "integer",
            "title": "Drafts Left"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          }
        },
        "type": "object",
        "required": [
          "draft",
          "drafts_left"
        ],
        "title": "DraftResponse",
        "description": "POST /games/{id}/draft response: a line the player may send as-is, edit, or drop.\nNothing has entered the game; POST /turns still says it."
      },
...
      "FundGame": {
        "properties": {
          "api_key": {
            "type": "string",
            "minLength": 1,
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "FundGame",
        "description": "POST /games/{id}/key body: a seat holder's key to resume a game that lost its\nfunding in a restart. Tried on the provider before the game moves; never stored."
      },
      "GameCreated": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "seat_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seat Token"
          }
        },
        "type": "object",
        "required": [
          "game_id"
        ],
        "title": "GameCreated",
        "description": "POST /games and POST /games/{id}/start response."
      },
...
      "GameStatus": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "waiting",
              "running",
              "finished",
              "dropped"
            ],
            "title": "State"
          },
          "server_time": {
            "type": "string",
            "title": "Server Time"
          },
          "players": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Players"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "max_seats": {
            "type": "integer",
            "title": "Max Seats",
            "default": 0
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "locked": {
            "type": "boolean",
            "title": "Locked",
            "default": false
          },
          "human_players": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Human Players"
          },
          "you": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "You"
          },
          "pending_input": {
            "type": "boolean",
            "title": "Pending Input",
            "default": false
          },
          "pending_seats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pending Seats"
          },
          "deadlines": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Deadlines"
          },
          "game_over": {
            "type": "boolean",
            "title": "Game Over",
            "default": false
          },
          "last_seq": {
            "type": "integer",
            "title": "Last Seq",
            "default": 0
          },
          "awaiting_key": {
            "type": "boolean",
            "title": "Awaiting Key",
            "default": false
          },
          "winner": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "villagers",
                  "wolves",
                  "serial_killer"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          },
          "alive_role_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Alive Role Counts"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "state",
          "server_time"
        ],
        "title": "GameStatus",
        "description": "GET /games/{id} response: the poll-side snapshot (the stream carries the rest).\n\nOne shape for both registry phases \u2014 a lobby fills only the header trio\n(game_id, state, players) and leaves the game fields at their defaults."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NewSoloGame": {
        "properties": {
          "human": {
            "type": "boolean",
            "title": "Human",
            "default": false
          },
          "human_role": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "villager",
                  "wolf",
                  "investigator",
                  "healer",
                  "serial_killer",
                  "vigilante"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Role"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "memory": {
            "type": "boolean",
            "title": "Memory",
            "default": false
          }
        },
        "type": "object",
        "title": "NewSoloGame",
        "description": "POST /games body: the solo door. One person's table, started on the spot, with\nthem in a seat or only watching an all-AI game. Rooms, where several people share a\ntable, are the other door (POST /rooms) and a deliberately separate contract."
      },
...
      "TurnAccepted": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "accepted"
          }
        },
        "type": "object",
        "title": "TurnAccepted",
        "description": "POST /games/{id}/turns response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}